Upgrade jhipster to 2.8.2
We upgraded the application to jhipster 2.8.2.
We ignored the output of jhipster-upgrade for the
java and typescript source code, since the suggested
code changes broke everything (even the SQL migration
XML scripts).

We had to remove the old metric system, but jhipster
now comes with a default metric system which should be
used instead.

Author: NiHa, LuGe